+++ src/main/java/ch/quantasy/tinkerforge/device/rotaryEncoder/RotaryEncoderDeviceCallback.java
@@ -58,6 +58,6 @@ public interface RotaryEncoderDeviceCallback extends DeviceCallback, BrickletRot
 
     public void countCallbackThresholdChanged(DeviceCountCallbackThreshold threshold);
 
-    public void countReset(long latestCount);
+    public void countReset(int latestCount);
 
 }
